use the same class for both symmetry types
